currency_consumer/requirements.txt
aiokafka
dependency_injector==4.46.0
clickhouse-sqlalchemy
